+++ pcs/cli/booth/console_report.py
@@ -5,9 +5,25 @@ from __future__ import (
 )
 
 from pcs.common import report_codes as codes
+from pcs.cli.common.console_report import (
+    format_list,
+    format_optional,
+)
 
 def format_booth_default(value, template):
     return "" if value in ("booth", "", None) else template.format(value)
+
+def booth_config_accepted_by_node(info):
+    desc = ""
+    if info["name_list"] and info["name_list"] not in [["booth"]]:
+        desc = "{_s} {_list}".format(
+            _s=("s" if len(info["name_list"]) > 1 else ""),
+            _list=format_list(info["name_list"])
+        )
+    return "{_node_info}Booth config{_desc} saved".format(
+        _node_info=format_optional(info["node"], "{0}: "),
+        _desc=desc,
+    )
 
 #Each value (a callable taking report_item.info) returns a message.
 #Force text will be appended if necessary.
@@ -80,26 +96,18 @@ CODE_TO_MESSAGE_BUILDER_MAP = {
         "Sending booth configuration to cluster nodes..."
     ,
 
-    codes.BOOTH_CONFIG_ACCEPTED_BY_NODE: lambda info:
-        "{node_info}Booth config{desc} saved.".format(
-            desc=(
-                "" if info["name_list"] in [None, [], ["booth"]]
-                else "(s) ({0})".format(", ".join(info["name_list"]))
-            ),
-            node_info="{0}: ".format(info["node"]) if info["node"] else ""
-        )
-    ,
+    codes.BOOTH_CONFIG_ACCEPTED_BY_NODE: booth_config_accepted_by_node,
 
     codes.BOOTH_CONFIG_DISTRIBUTION_NODE_ERROR: lambda info:
-        "Unable to save booth config{desc} on node '{node}': {reason}".format(
-            desc=format_booth_default(info["name"], " ({0})"),
+        "Unable to save booth config{_desc} on node '{node}': {reason}".format(
+            _desc=format_booth_default(info["name"], " '{0}'"),
             **info
         )
     ,
 
     codes.BOOTH_CONFIG_READ_ERROR: lambda info:
-        "Unable to read booth config{desc}.".format(
-            desc=format_booth_default(info["name"], " ({0})")
+        "Unable to read booth config{_desc}".format(
+            _desc=format_booth_default(info["name"], " '{0}'")
         )
     ,
 
@@ -136,5 +144,9 @@ CODE_TO_MESSAGE_BUILDER_MAP = {
 
     codes.BOOTH_CANNOT_IDENTIFY_KEYFILE:
         "cannot identify authfile in booth configuration"
+    ,
+
+    codes.BOOTH_UNSUPPORTED_FILE_LOCATION: lambda info:
+        "Path '{file}' is not supported for booth config files".format(**info)
     ,
 }