+++ pcs/lib/booth/test/test_sync.py
@@ -276,7 +276,7 @@ class SendAllConfigToNodeTest(TestCase):
                     {
                         "node": self.node.label,
                         "name_list": [
-                            "name1.conf", "file1.key", "name2.conf", "file2.key"
+                            "file1.key", "file2.key", "name1.conf", "name2.conf"
                         ]
                     }
                 )
@@ -450,7 +450,7 @@ class SendAllConfigToNodeTest(TestCase):
                     report_codes.BOOTH_CONFIG_ACCEPTED_BY_NODE,
                     {
                         "node": self.node.label,
-                        "name_list": ["name2.conf", "file2.key"]
+                        "name_list": ["file2.key", "name2.conf"]
                     }
                 )
             ]
@@ -860,7 +860,7 @@ class SendAllConfigToNodeTest(TestCase):
                     report_codes.BOOTH_CONFIG_ACCEPTED_BY_NODE,
                     {
                         "node": self.node.label,
-                        "name_list": ["name1.conf", "name2.conf", "file2.key"]
+                        "name_list": ["file2.key", "name1.conf", "name2.conf"]
                     }
                 )
             ]
@@ -943,7 +943,7 @@ class SendAllConfigToNodeTest(TestCase):
                     report_codes.BOOTH_CONFIG_ACCEPTED_BY_NODE,
                     {
                         "node": self.node.label,
-                        "name_list": ["name2.conf", "file2.key"]
+                        "name_list": ["file2.key", "name2.conf"]
                     }
                 )
             ]